move IsEqualIgnoreCaseAndWhiteSpace to MatcherUtils
