+++ src/test/java/com/amazon/opendistroforelasticsearch/sql/util/MatcherUtils.java
@@ -195,4 +195,48 @@ public class MatcherUtils {
             }
         };
     }
+
+
+    /**
+     * Tests if a string is equal to another string, ignore the case and whitespace.
+     */
+    public static class IsEqualIgnoreCaseAndWhiteSpace extends TypeSafeMatcher<String> {
+        private final String string;
+
+        public IsEqualIgnoreCaseAndWhiteSpace(String string) {
+            if (string == null) {
+                throw new IllegalArgumentException("Non-null value required");
+            }
+            this.string = string;
+        }
+
+        @Override
+        public boolean matchesSafely(String item) {
+            return ignoreCase(ignoreSpaces(string)).equals(ignoreCase(ignoreSpaces(item)));
+        }
+
+        @Override
+        public void describeMismatchSafely(String item, Description mismatchDescription) {
+            mismatchDescription.appendText("was ").appendValue(item);
+        }
+
+        @Override
+        public void describeTo(Description description) {
+            description.appendText("a string equal to ")
+                    .appendValue(string)
+                    .appendText(" ignore case and white space");
+        }
+
+        public String ignoreSpaces(String toBeStripped) {
+            return toBeStripped.replaceAll("\\s+", "").trim();
+        }
+
+        public String ignoreCase(String toBeLower) {
+            return toBeLower.toLowerCase();
+        }
+
+        public static Matcher<String> equalToIgnoreCaseAndWhiteSpace(String expectedString) {
+            return new IsEqualIgnoreCaseAndWhiteSpace(expectedString);
+        }
+    }
 }
